+++ doc/src/release-notes.md
@@ -13,8 +13,13 @@
 * `[hoomd-mc`]: Add `InsertReMoveCount` type that counts the accepted and rejected insertion and removal moves (#354).
 * `[hoomd-microstate`]: Implement all traits necessary to use `Hyperparallelepiped`, `Triclinic`, and `Rhomboid` as simulation boundaries (#88).
 * `[hoomd-microstate`]: Implement `AppendMicrostate` for microstates with `Triclinic` and `Rhomboid` boundaries (#88).
+* `[hoomd-microstate`]: Add `Microstate::iter_bodies_tag_order` method that iterates over all bodies in tag order (#393).
+* `[hoomd-microstate`]: Add `Replicate` trait and implement it for microstates with periodic `Rhomboid`, `Triclinic`, and `Hypercuboid` boundaries.
+   Call `microstate.replicate` to repeat a periodic structure N x N (...) x N times (#393).
 * `[hoomd-vector`]: Implement `Mul<Cartesian<N>> for f64`, `Mul<PositiveReal> for Cartesian<N>`, `Mul<Cartesian<N>> for PositiveReal`, `MulAssign<f64> for Cartesian<N>`, and `MulAssign<PositiveReal> for Cartesian<N>` (#88).
+* `[hoomd-vector`]: Add `Cartesian::basis` method that constructs Cartesian basis vectors (#393).
 * `[hoomd-workspace`]: Add the `hoomd-workspace` crate. Use it to initialize and read state points in a signac workspace (#359).
+* *examples*: Add "Melt an ideal hexagonal structure of hexagons" (#393).
 
 *Changed:*
 
@@ -29,6 +34,7 @@
 * `[hoomd-interaction]`: Fix typos in documentation (#358).
 * `[hoomd-microstate]`: Fix typos in documentation (#358).
 * *examples*: Properly auto size egui parameter windows (#354).
+* *examples*: Reduce the zoom speed (#393).
 
 ## 1.2.0 (2026-07-07)